# internal/modules/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 九小二 Backend API
  version: 1.0.0-cp1
  description: 九小二产品一期 Go 后端唯一公开契约，覆盖交易、履约、打印、通知、账号运营与酒类合规。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/rider-application/sms-login:
    post:
      summary: 待审核骑手使用手机号和短信验证码登录并签发受限 Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SmsLoginReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /rider-applications:
    post:
      summary: 骑手首次提交申请
      description: 相同 Idempotency-Key 与相同请求返回首次结果且不重复消费短信验证码；新的幂等键必须使用尚未消费的一次性验证码。
      security: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RiderApplicationSubmitReq" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /rider-applications/me:
    get:
      summary: 查看本人骑手申请
      security:
        - applicationBearer: []
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      summary: 修改被拒绝的本人申请
      security:
        - applicationBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RiderApplicationUpdateReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /rider-applications/me/resubmit:
    post:
      summary: 重新提交原骑手申请
      security:
        - applicationBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RiderApplicationVersionReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
...
  /realtime/tickets:
    post:
      summary: 签发骑手 WebSocket 一次性连接票据
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RealtimeTicketReq" }
      responses:
        "201": { $ref: "#/components/responses/RealtimeTicket" }
        default: { $ref: "#/components/responses/Problem" }
  /realtime/ws:
    get:
      summary: 骑手实时履约 WebSocket（协议版本 1）
      description: 使用一次性 ticket 完成升级；升级后支持 hello、resume、event、ack、ack_result、sync_complete、error 与 server_shutdown 帧。
      security: []
      parameters:
        - name: ticket
          in: query
          required: true
          schema: { type: string, minLength: 20, maxLength: 256 }
      responses:
        "101": { description: WebSocket upgrade accepted }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: application_access JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
  responses:
    Success:
      description: 请求成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
...
    RealtimeTicket:
      description: 一次性 WebSocket 票据签发成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RealtimeTicketEnvelope" }
...
    Problem:
      description: 标准业务或系统错误
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  schemas:
...
    RealtimeTicketReq:
      type: object
      additionalProperties: false
      required: [device_id, platform, client_version, protocol_version]
      properties:
        device_id: { type: string, minLength: 1, maxLength: 128 }
        platform: { type: string, enum: [weapp, h5, android, ios, test] }
        client_version: { type: string, minLength: 1, maxLength: 32 }
        protocol_version: { type: integer, enum: [1] }
...
    SuccessEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { type: object, additionalProperties: true, nullable: true }
        request_id: { type: string }
    RealtimeTicketEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          additionalProperties: false
          required: [ticket, expires_at, ws_path, heartbeat_interval_seconds, max_resume_items, protocol_version]
          properties:
            ticket: { type: string, minLength: 20, writeOnly: true }
            expires_at: { type: string, format: date-time }
            ws_path: { type: string, example: /api/v1/realtime/ws }
            heartbeat_interval_seconds: { type: integer, minimum: 5 }
            max_resume_items: { type: integer, minimum: 1, maximum: 500 }
            protocol_version: { type: integer, enum: [1] }
        request_id: { type: string }
...
    SmsLoginReq:
      type: object
      required: [phone, code]
      properties:
        phone: { type: string }
        code: { type: string }
...
    RiderApplicationServiceScope:
      type: object
      additionalProperties: false
      required: [shop_ids]
      properties:
        shop_ids:
          type: array
          minItems: 1
          maxItems: 50
          uniqueItems: true
          items: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
    RiderApplicationSubmitReq:
      type: object
      additionalProperties: false
      required: [name, phone, code, service_scope]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        phone: { type: string, pattern: "^1[3-9][0-9]{9}$" }
        code: { type: string, pattern: "^[0-9]{6}$" }
        service_scope: { $ref: "#/components/schemas/RiderApplicationServiceScope" }
    RiderApplicationUpdateReq:
      type: object
      additionalProperties: false
      required: [name, service_scope, expected_version]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        service_scope: { $ref: "#/components/schemas/RiderApplicationServiceScope" }
        expected_version: { type: integer, minimum: 1 }
    RiderApplicationVersionReq:
      type: object
      additionalProperties: false
      required: [expected_version]
      properties:
        expected_version: { type: integer, minimum: 1 }
...
    ProblemDetails:
      type: object
      required: [type, title, status, error_code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error_code:
          type: string
          description: 幂等写请求同 key 异参固定返回 IDEMPOTENCY_KEY_REUSED；同参请求仍处于有效处理租约时固定返回 IDEMPOTENCY_IN_PROGRESS。
        request_id: { type: string }